+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 * bug-fix: Fix FileNotFoundError for entry_point without source_dir
 * doc-fix: Add missing feature 1.5.0 in change log
 * doc-fix: Add README for airflow
+* feature: Add Amazon SageMaker Semantic Segmentation algorithm to the registry.
 
 1.15.1
 ======